requirements.txt
discord.py[voice]>=2.3
topggpy>=1.4
python-dotenv>=1.0
jishaku>=2.5
